intelligence-layer/osint-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
httpx==0.27.2
python-whois==0.9.4
dnspython==2.7.0
python-dotenv==1.0.1
